updated cli
