+++ toolchain/exec.go
@@ -8,6 +8,14 @@ import (
 
 var execFunc = syscall.Exec
 
+// ToolRunner defines the interface for running and resolving tools (for real and mock installers)
+type ToolRunner interface {
+	FindBinaryPath(owner, repo, version string) (string, error)
+	GetResolver() ToolResolver
+	CreateLatestFile(owner, repo, version string) error
+	ReadLatestFile(owner, repo string) (string, error)
+}
+
 // RunExecCommand contains business logic for executing tools.
 // It does not depend on cobra.Command, only raw args.
 func RunExecCommand(installer ToolRunner, args []string) error {
